getting task by user authenticated

+++ backend/src/controllers/ProjectController.ts
@@ -6,10 +6,10 @@ export class ProjectController {
   // Método estático para crear proyectos
   static createProject = async (req: Request, res: Response) => {
     const project = new Project(req.body);
-
-    console.log(req.user);
     
-
+    //Asigna un manager
+    project.manager = req.user.id
+    
     try {
       await project.save();
       res.send("Proyecto creado correctamente");
@@ -21,7 +21,11 @@ export class ProjectController {
   // Método estático para obtener  proyectos
   static getAllProjects = async (req: Request, res: Response) => {
     try {
-      const projects = await Project.find({});
+      const projects = await Project.find({
+        $or: [
+          {manager: {$in: req.user.id}}
+        ]
+      });
       res.json(projects);
     } catch (error) {
       console.log(error);
@@ -38,6 +42,11 @@ export class ProjectController {
         const error = new Error("Proyecto no encontrado");
         return res.status(404).json({ error: error.message });
       }
+      
+      if(project.manager.toString() !== req.user.id.toString()){
+        const error = new Error("Accion no valida");
+        return res.status(404).json({ error: error.message });  
+      }
       res.json(project);
     } catch (error) {
       console.log(error);
@@ -53,6 +62,10 @@ export class ProjectController {
       if (!project) {
         const error = new Error("Proyecto no encontrado");
         return res.status(404).json({ error: error.message });
+      }
+      if(project.manager.toString() !== req.user.id.toString()){
+        const error = new Error("Solo el manager puede realizar esta accion");
+        return res.status(404).json({ error: error.message });  
       }
       project.clientName = req.body.clientName;
       project.projectName = req.body.projectName;
@@ -74,6 +87,10 @@ export class ProjectController {
       if (!project) {
         const error = new Error("Proyecto no encontrado");
         return res.status(404).json({ error: error.message });
+      }
+      if(project.manager.toString() !== req.user.id.toString()){
+        const error = new Error("Solo el manager puede realizar esta accion");
+        return res.status(404).json({ error: error.message });  
       }
 
       await project.deleteOne();
